Add tests for reliable ordering and delegate invocation
If two methods are called on an actor one after the other by a single actor, no
guarantees are currently made that they will end up being called in the same
order. This makes it very hard to correctly code actors and introduces subtle
bugs when people forget to take this into account.

Thus the behaviour is going to change to enforce ordering in some circumstances,
this test checks for the correct sequential behaviour (and thus currently fails)